# rest-api/docs/swagger/swagger.yaml
basePath: /
definitions:
  github_com_user_simple-blog_internal_modules_auth_domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_user_simple-blog_internal_modules_auth_domain.LoginResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  github_com_user_simple-blog_internal_modules_auth_domain.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  github_com_user_simple-blog_internal_modules_auth_domain.RegisterResponse:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  github_com_user_simple-blog_internal_modules_comments_domain.CreateCommentRequest:
    properties:
      content:
        minLength: 1
        type: string
    required:
    - content
    type: object
  github_com_user_simple-blog_internal_modules_comments_domain.UpdateCommentRequest:
    properties:
      content:
        minLength: 1
        type: string
    required:
    - content
    type: object
  github_com_user_simple-blog_internal_modules_health_domain.HealthResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  github_com_user_simple-blog_internal_modules_posts_domain.CreatePostRequest:
    properties:
      content:
        type: string
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - content
    - title
    type: object
  github_com_user_simple-blog_internal_modules_posts_domain.PaginatedPostResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Post'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_user_simple-blog_internal_modules_posts_domain.UpdatePostRequest:
    properties:
      content:
        type: string
      title:
        maxLength: 255
        minLength: 3
        type: string
    type: object
  github_com_user_simple-blog_internal_platform_errors.AppError:
    properties:
      code:
        $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_user_simple-blog_internal_platform_errors.ErrorCode:
    enum:
    - INTERNAL_ERROR
    - NOT_FOUND
    - BAD_REQUEST
    - UNAUTHORIZED
    - FORBIDDEN
    - UNPROCESSABLE_ENTITY
    - CONFLICT
    type: string
    x-enum-varnames:
    - CodeInternal
    - CodeNotFound
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeUnprocessable
    - CodeConflict
  github_com_user_simple-blog_internal_platform_errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.AppError'
      success:
        type: boolean
    type: object
  github_com_user_simple-blog_internal_platform_response.Response:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
  models.Comment:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      post_id:
        type: string
      updated_at:
        type: string
    type: object
  models.Post:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is a sample blog system server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Blog System API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_user_simple-blog_internal_modules_auth_domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_user_simple-blog_internal_platform_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_user_simple-blog_internal_modules_auth_domain.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
      summary: Login user
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account and assign default role
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_user_simple-blog_internal_modules_auth_domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_user_simple-blog_internal_platform_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_user_simple-blog_internal_modules_auth_domain.RegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
  /comments/{id}:
    delete:
      description: Delete a comment by ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_response.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update an existing comment by ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_user_simple-blog_internal_modules_comments_domain.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_user_simple-blog_internal_platform_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Comment'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update a comment
      tags:
      - Comments
  /health:
    get:
      consumes:
      - application/json
      description: Get the status of the service and database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_modules_health_domain.HealthResponse'
      summary: Check service health
      tags:
      - Health
  /posts:
    get:
      description: Get paginated list of posts
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_user_simple-blog_internal_platform_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_user_simple-blog_internal_modules_posts_domain.PaginatedPostResponse'
              type: object
      summary: Get all posts
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: Create a new blog post
      parameters:
      - description: Post details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_user_simple-blog_internal_modules_posts_domain.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_user_simple-blog_internal_platform_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Post'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new post
      tags:
      - Posts
  /posts/{id}:
    delete:
      description: Delete a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a post
      tags:
      - Posts
    get:
      description: Get a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_user_simple-blog_internal_platform_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Post'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
      summary: Get post by ID
      tags:
      - Posts
    put:
      consumes:
      - application/json
      description: Update an existing post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_user_simple-blog_internal_modules_posts_domain.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_user_simple-blog_internal_platform_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Post'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_user_simple-blog_internal_platform_errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a post
      tags:
      - Posts
  /posts/{id}/comments:
    post:
      consumes:
      - application/json
      description: Add a new comment to an existing blog post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_user_simple-blog_internal_modules_comments_domain.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_user_simple-blog_internal_platform_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Comment'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Add a comment to a post
      tags:
      - Comments
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"